+++ src/bin/digital_intraday.rs
@@ -32,13 +32,14 @@
 use std::process;
 
 use alpha_vantage_rust::{
-  alpha_lib::{alpha_io::base::load_intraday, misc_functions::get_exe_name},
-  db_funcs::{get_proc_id_or_insert, get_symbols_and_sids_for, log_proc_end, log_proc_start},
-  dbfunctions::base::establish_connection_or_exit,
-  security_types::sec_types::SecurityType,
+    alpha_lib::{alpha_io::base::load_intraday, misc_functions::get_exe_name},
+    db_funcs::{get_proc_id_or_insert, log_proc_end, log_proc_start},
+    dbfunctions::base::establish_connection_or_exit,
+    security_types::sec_types::SecurityType,
 };
 use dotenvy::dotenv;
 use indicatif::ProgressBar;
+use alpha_vantage_rust::dbfunctions::symbols::get_symbols_and_sids_for;
 
 fn main() {
   dotenv().ok();
