+++ src/dbfunctions.rs
@@ -36,6 +36,10 @@ pub mod feed;
 
 pub mod raw_queries;
 pub mod sources;
+pub  mod symbols;
 pub mod ticker_sentiments;
 pub mod topic_maps;
 pub mod topic_refs;
+pub mod overview;
+pub mod price;
+pub mod combined;
